added login .js and .vue files
