+++ App/Client/Views/login.vue
@@ -1,0 +1,8 @@
+<template>
+  <div class="login">
+    <form class="login-form">
+      <input placeholder="Enter your username here" type="text" v-model="username">
+      <button class="login-button" role="button" href="#" @:click.prevent="login"> Log In </button>
+    </form>
+  </div>
+</template>
