--- App/Client/login.js
@@ -1,7 +1,0 @@
-import Vue from 'vue';
-import Login from './login.vue'
-
-var login = new Vue({
-  el: '.login',
-  components: { Login }
-})
