+++ package.json
@@ -53,6 +53,7 @@
     "vue-hot-reload-api": "^1.2.0",
     "vue-html-loader": "^1.2.3",
     "vue-loader": "^7.5.3",
+    "vue-resource": "^1.0.3",
     "vue-router": "^2.1.1",
     "webpack": "^1.13.3",
     "webpack-stream": "^3.2.0",
